cloud-infra/project1/src/requirements.txt
# Function dependencies, for example:
# package>=version
click==7.1.2
cloudevents==1.2.0
deprecation==2.1.0
Flask==1.1.2
functions-framework==2.2.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathtools==0.1.2
watchdog==1.0.2
Werkzeug==1.0.1
google-cloud-compute
google-cloud-pubsub==2.9.0
google-api-python-client